Tell the bot owner when starve_check should probably be executed after reloading
If the bot is reloaded right before a starve_check was going to happen, then the entire interval's worth of time is lost. In this situation, starve_check should probably be run after the bot is loaded back to move the bot to the next loop. The ratio 0.04 was chosen because it would trigger the warning if the bot was reloaded within 1 hour before the starve_check for a 24 hour starve_time interval.