Changes to formatting, organization, bug fixes
It is no longer possible to submit an answer while scores are being assigned to possibly get a score of less than 500 for the question.

The end of game leaderboard's formatting will no longer break if ranks go to two digits. It will also note when players will not earn any credits because they do not have a bank account, instead of raising a NoAccount error.