Change all upgrade abbreviations to full name
This breaks upgrade_menu and activate_menu, which are about to be removed anyways. This eliminates the need to convert between the abbreviated dictionary keys and the full name at various points in the program. The long keys look ugly for now, but they may be changed later if a better solution can be found.